+++ benchmarks/filters.rb
@@ -15,24 +15,25 @@ filters = ActiveInteraction::Filter
   .send(:const_get, :CLASSES)
   .reject { |_, v| v.name =~ /\bAbstract/ }
 
-# rubocop:disable all
 VALUES = {
-  array:     [Array.new], # TODO
+  array:     [[]], # TODO
   boolean:   [false, '0', 'false', true, '1', 'true'],
   date:      [Date.new, Date.new.to_s],
   date_time: [DateTime.new, DateTime.new.to_s],
   decimal:   [BigDecimal.new(0)],
-  file:      [File.new(__FILE__), Struct.new(:tempfile).new(File.new(__FILE__))],
+  file:      [
+    File.new(__FILE__),
+    Struct.new(:tempfile).new(File.new(__FILE__))
+  ],
   float:     [0.0, '0.0', 0],
-  hash:      [Hash.new], # TODO
+  hash:      [{}], # TODO
   integer:   [0, '0', 0.0],
   interface: [Object.new],
-  model:     [Object.new], # TODO: Reconstantizing.
+  object:    [Object.new], # TODO: Reconstantizing.
   string:    [''], # TODO: Without strip.
   symbol:    [:'', ''],
   time:      [Time.at(0), Time.at(0).to_s, 0] # TODO: TimeWithZone
 }
-# rubocop:enable all
 
 missing_filters = filters.keys - VALUES.keys
 fail "Missing filters: #{missing_filters}" unless missing_filters.empty?