+++ example/app/interactions/things/list_things.rb
@@ -1,0 +1,10 @@
+# coding: utf-8
+
+require 'active_interaction'
+
+# List all the things.
+class ListThings < ActiveInteraction::Base
+  def execute
+    Thing.all
+  end
+end