+++ example/config/environment.rb
@@ -1,0 +1,5 @@
+# coding: utf-8
+
+require_relative 'application'
+
+Rails.application.initialize!