+++ example/config/routes.rb
@@ -1,0 +1,5 @@
+# coding: utf-8
+
+Rails.application.routes.draw do
+  resources :things
+end