+++ spec/active_interaction/integration/hash_interaction_spec.rb
@@ -27,6 +27,11 @@ describe HashInteraction do
     it 'returns the correct value for :b' do
       expect(result[:b]).to eql('x' => {})
     end
+
+    it 'does not raise an error with an invalid nested value' do
+      inputs[:a] = { x: false }
+      expect { outcome }.to_not raise_error
+    end
   end
 
   context 'with an invalid default' do
@@ -39,11 +44,31 @@ describe HashInteraction do
     end
   end
 
+  context 'with an invalid default as a proc' do
+    it 'does not raise an error' do
+      expect do
+        Class.new(ActiveInteraction::Base) do
+          array :a, default: -> { Object.new }
+        end
+      end.to_not raise_error
+    end
+  end
+
   context 'with an invalid nested default' do
-    it 'raises an error' do
+    it 'raises an error with a non-empty hash' do
       expect do
         Class.new(ActiveInteraction::Base) do
           hash :a, default: { x: Object.new } do
+            hash :x
+          end
+        end
+      end.to raise_error ActiveInteraction::InvalidDefaultError
+    end
+
+    it 'raises an error' do
+      expect do
+        Class.new(ActiveInteraction::Base) do
+          hash :a, default: {} do
             hash :x
           end
         end