+++ spec/active_interaction/integration/model_interaction_spec.rb
@@ -1,7 +1,0 @@
-# coding: utf-8
-
-require 'spec_helper'
-
-describe 'ModelInteraction' do
-  it_behaves_like 'an interaction', :model, -> { // }, class: Regexp
-end